fix(planner): restore Pagar action + guard useless long-press
- ExpenseCardDraggable: add onPay handler + 'Pagar' as first dropdown item
  (hidden when status === COMPLETED). Restores the Pagar/Editar/Eliminar
  action surface lost in the drag-drop migration.
- DragEnvelopeBoard: mount PayExpenseDialog + manage payTarget state; widen
  accounts type to PlanAccount so it matches the dialog's shape.
- MobilePeriodoView: drop dead editTarget/payTarget state + dialogs (now
  owned by the board).
- Long-press guard: skip activation when remainder <= 0 to prevent the
  overlay from showing '$0 sin asignar' + 'cubre $0' chips on
  already-fully-assigned expenses.
